Endpoint de registro de usuarios Con Validaciones OK
