+++ src/main/java/com/alkemy/paypocket/dtos/UserDto.java
@@ -1,5 +1,7 @@
 package com.alkemy.paypocket.dtos;
 
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -11,12 +13,16 @@ import lombok.Setter;
 @NoArgsConstructor
 public class UserDto {
 
+    @NotBlank(message = "Campo firstName OBLIGATORIO ")
     private String firstName;
 
+    @NotBlank(message = "Campo lastName OBLIGATORIO ")
     private String lastName;
 
+    @NotBlank(message = "Campo email OBLIGATORIO ")
     private String email;
 
+    @NotBlank(message = "Campo passwords OBLIGATORIO ")
     private String passwords;
 
     private String rolName;
