+++ src/main/java/com/alkemy/paypocket/services/UserService.java
@@ -5,7 +5,6 @@ import com.alkemy.paypocket.entities.User;
 import com.alkemy.paypocket.mappers.UserMapper;
 import com.alkemy.paypocket.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -23,7 +22,6 @@ public class UserService {
         userRepository.save(newUser);
 
         return newUser;
-
     }
